requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
torch==2.0.0
torchvision==0.15.0
pillow==9.5.0
transformers==4.30.2
numpy==1.24.3
huggingface-hub==0.16.4
timm==0.4.12
sentence-transformers==2.2.2
weaviate-client==4.4.3
dependency-injector==4.41.0
loguru==0.7.2
tqdm==4.66.1
starlette==0.27.0
pydantic==2.4.2
typing-extensions==4.8.0